+++ include/CGAL/Approximate_min_ellipsoid_d/Khachiyan_approximation.h
@@ -11,8 +11,8 @@
 // This file is provided AS IS with NO WARRANTY OF ANY KIND, INCLUDING THE
 // WARRANTY OF DESIGN, MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE.
 //
-// $URL: svn+ssh://scm.gforge.inria.fr/svn/cgal/branches/CGAL-3.7-branch/Approximate_min_ellipsoid_d/include/CGAL/Approximate_min_ellipsoid_d/Khachiyan_approximation.h $
-// $Id: Khachiyan_approximation.h 36397 2007-02-16 15:56:11Z gaertner $
+// $URL: svn+ssh://scm.gforge.inria.fr/svn/cgal/trunk/Approximate_min_ellipsoid_d/include/CGAL/Approximate_min_ellipsoid_d/Khachiyan_approximation.h $
+// $Id: Khachiyan_approximation.h 58840 2010-09-23 19:23:50Z afabri $
 // 
 //
 // Author(s)     : Kaspar Fischer <fischerk@inf.ethz.ch>
@@ -336,7 +336,7 @@ namespace CGAL {
       const int n_old = n;
       for (; first != last; first++)
 	P.push_back(&*first);
-      n = P.size();
+      n = static_cast<int>(P.size());
 
       // debugging output:
       CGAL_APPEL_LOG("appel","  Add()'ing " << n-n_old << " points to a " <<
